Create B2BStorefrontTokenService
