+++ packages/core/src/checkout/create-checkout-service.ts
@@ -10,6 +10,7 @@ import { getDefaultLogger } from '../common/log';
 import { getEnvironment } from '../common/utility';
 import {
     B2BStorefrontTokenRequestSender,
+    B2BStorefrontTokenService,
     CompanyAddressRequestSender,
     CompanyAddressService,
 } from '../company';
@@ -249,8 +250,10 @@ export default function createCheckoutService(options?: CheckoutServiceOptions):
             new B2BTokenRequestSender(requestSender),
         ),
         new CompanyAddressService(
-            store,
-            new B2BStorefrontTokenRequestSender(requestSender),
+            new B2BStorefrontTokenService(
+                store,
+                new B2BStorefrontTokenRequestSender(requestSender),
+            ),
             new CompanyAddressRequestSender(new GraphQLRequestSender(requestSender)),
         ),
     );
