+++ packages/core/src/company/index.ts
@@ -1,4 +1,5 @@
 export { default as B2BStorefrontTokenRequestSender } from './b2b-storefront-token-request-sender';
+export { default as B2BStorefrontTokenService } from './b2b-storefront-token-service';
 export { default as CompanyAddressRequestSender } from './company-address-request-sender';
 export { default as CompanyAddressService } from './company-address-service';
 export {
